Provisional changes

+++ src/main/java/ar/edu/um/comidar/controller/RestaurantController.java
@@ -75,7 +75,9 @@ public class RestaurantController {
 			restaurant.setLastUpdateDate(timestamp);
 			restaurantService.create(restaurant);
 			restaurant.setImageUrl("/restaurants/" + restaurant.getRestaurantId() + "." + restaurant.getRestaurantImage().getExtension());
-			imageService.uploadImage(restaurant.getRestaurantImage().getFile().getInputStream(), restaurant.getImageUrl());
+			if(!restaurant.getRestaurantImage().getFile().isEmpty()) {
+				imageService.uploadImage(restaurant.getRestaurantImage().getFile().getInputStream(), restaurant.getImageUrl());
+			}
 			restaurant.setRestaurantId(restaurant.getRestaurantId());
 			restaurantService.update(restaurant);
 			redirectAttributes.addFlashAttribute("message","Actualizacion realizada");
@@ -106,8 +108,10 @@ public class RestaurantController {
 			restaurant.setImageUrl("/restaurants/" + restaurant.getRestaurantId() + "." + restaurant.getRestaurantImage().getExtension());
 			restaurant.setCreationDate(restaurantService.findById(restaurant.getRestaurantId()).getCreationDate());
 			restaurant.setLastUpdateDate(timestamp);
-			imageService.deleteImage(restaurant.getImageUrl());
-			imageService.uploadImage(restaurant.getRestaurantImage().getFile().getInputStream(), restaurant.getImageUrl());
+			if(!restaurant.getRestaurantImage().getFile().isEmpty()) {
+				imageService.deleteImage(restaurant.getImageUrl());
+				imageService.uploadImage(restaurant.getRestaurantImage().getFile().getInputStream(), restaurant.getImageUrl());
+			}
 			restaurantService.update(restaurant);
 			redirectAttributes.addFlashAttribute("message","Actualizacion realizada");
 			redirectAttributes.addFlashAttribute("css","alert-success");
@@ -124,7 +128,6 @@ public class RestaurantController {
 	@GetMapping("/delete")
 	public String deleteRestaurant(@RequestParam(value="id",required=true) Long id, final RedirectAttributes redirectAttributes) throws UploadErrorException, DbxException, IOException{
 		Restaurant restaurant = restaurantService.findById(id);
-		imageService.deleteImage(restaurant.getImageUrl());
 		restaurantService.remove(restaurant);
 		redirectAttributes.addFlashAttribute("message","Se ha eliminado el usuario exitosamente");
 		redirectAttributes.addFlashAttribute("css","alert-success");
@@ -133,8 +136,12 @@ public class RestaurantController {
 	}
 
 	@GetMapping("/list/all")
-	public ResponseEntity<List<Restaurant>> sendRestaurantList(){
-		return new ResponseEntity<>(restaurantService.findAll(),HttpStatus.OK);
+	public ResponseEntity<List<Restaurant>> sendRestaurantList() throws UploadErrorException, DbxException, IOException{
+		List<Restaurant> restaurantList = restaurantService.findAll();
+		for (Restaurant restaurant : restaurantList) {
+			restaurant.setImageTemporaryUrl(imageService.getImageURL(restaurant.getImageUrl()));
+		}
+		return new ResponseEntity<>(restaurantList,HttpStatus.OK);
 	}
 	@GetMapping("/list/search")
 	public ResponseEntity<List<Restaurant>> sendRestaurantListSearch(){
